--- webpack.config.js
@@ -1,85 +1,0 @@
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const Dotenv = require('dotenv-webpack');
-const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
-
-const isProduction = process.env.NODE_ENV === 'production';
-
-const config = {
-  mode: isProduction ? 'production' : 'development',
-  entry: './src/index.tsx',
-  output: {
-    filename: 'bundle.js',
-    path: path.resolve(__dirname, 'dist'),
-    clean: true,
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: './public/index.html',
-      filename: 'index.html',
-      inject: 'body',
-    }),
-    new ForkTsCheckerWebpackPlugin({
-      async: false,
-    }),
-    ...(isProduction ? [new MiniCssExtractPlugin({ filename: '[name].[contenthash].css' })] : []),
-    new Dotenv({
-      path: `./.env.${process.env.NODE_ENV}`,
-    }),
-    !isProduction && new ReactRefreshWebpackPlugin(),
-  ].filter(Boolean),
-  module: {
-    rules: [
-      {
-        test: /\.(ts|tsx)$/i,
-        loader: 'ts-loader',
-        exclude: /node_modules/,
-      },
-      {
-        test: /\.(js|jsx|ts|tsx)$/,
-        use: [
-          {
-            loader: 'babel-loader',
-            options: {
-              plugins: [!isProduction && require.resolve('react-refresh/babel')].filter(Boolean),
-            },
-          },
-        ],
-        exclude: /node_modules/,
-      },
-      {
-        test: /\.(png|jpg|jpeg|gif|svg)$/i,
-        type: 'asset/resource',
-      },
-      {
-        test: /\.(woff|woff2|eot|ttf|otf)$/i,
-        type: 'asset/resource',
-      },
-      {
-        test: /\.css$/,
-        use: [isProduction ? MiniCssExtractPlugin.loader : 'style-loader', 'css-loader', 'postcss-loader'],
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.tsx', '.ts', '.jsx', '.js', '...'],
-  },
-  devtool: isProduction ? 'hidden-source-map' : 'eval-cheap-module-source-map',
-  devServer: {
-    static: {
-      directory: path.resolve(__dirname, 'public'),
-    },
-    port: 3000,
-    open: true,
-    hot: true,
-    compress: true,
-    historyApiFallback: true,
-    client: {
-      overlay: false,
-    },
-  },
-};
-
-module.exports = config;
