+++ webpack.prod.js
@@ -1,0 +1,28 @@
+// webpack.prod.js
+
+const { merge } = require('webpack-merge');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  mode: 'production', // 프로덕션 모드 설정
+  devtool: 'hidden-source-map',
+  plugins: [...[new MiniCssExtractPlugin({ filename: '[name].[contenthash].css' })]].filter(Boolean),
+  module: {
+    rules: [
+      {
+        test: /\.(js|jsx|ts|tsx)$/,
+        exclude: /node_modules/,
+        use: [
+          {
+            loader: 'babel-loader',
+          },
+        ],
+      },
+      {
+        test: /\.css$/,
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'],
+      },
+    ],
+  },
+});
